Flytter logikk som ikke skal brukes i prod til test
